+++ src/main/java/tz/go/moh/him/thscp/mediator/elmis/MediatorMain.java
@@ -10,15 +10,15 @@ import org.openhim.mediator.engine.MediatorServer;
 import org.openhim.mediator.engine.RegistrationConfig;
 import org.openhim.mediator.engine.RoutingTable;
 import org.openhim.mediator.engine.StartupActorsConfig;
-import tz.go.moh.him.thscp.mediator.elmis.Orchestration.EmergencyCommodityStockStatusOrchestrator;
-import tz.go.moh.him.thscp.mediator.elmis.Orchestration.ForecastAccuracyPerProgramOrchestrator;
-import tz.go.moh.him.thscp.mediator.elmis.Orchestration.LaboratoryDiagnosticEquipmentFunctionalityOrchestrator;
-import tz.go.moh.him.thscp.mediator.elmis.Orchestration.PercentageOfReportsAndRequisitionOrchestrator;
-import tz.go.moh.him.thscp.mediator.elmis.Orchestration.ReportingTimelinessOrchestrator;
-import tz.go.moh.him.thscp.mediator.elmis.Orchestration.StockAvailabilityOrchestrator;
-import tz.go.moh.him.thscp.mediator.elmis.Orchestration.StockOnHandStatusOrchestrator;
-import tz.go.moh.him.thscp.mediator.elmis.Orchestration.TurnAroundTimeOrchestrator;
-import tz.go.moh.him.thscp.mediator.elmis.Orchestration.WorkforceOrchestrator;
+import tz.go.moh.him.thscp.mediator.elmis.orchestrator.EmergencyCommodityStockStatusOrchestrator;
+import tz.go.moh.him.thscp.mediator.elmis.orchestrator.ForecastAccuracyPerProgramOrchestrator;
+import tz.go.moh.him.thscp.mediator.elmis.orchestrator.LaboratoryDiagnosticEquipmentFunctionalityOrchestrator;
+import tz.go.moh.him.thscp.mediator.elmis.orchestrator.PercentageOfReportsAndRequisitionOrchestrator;
+import tz.go.moh.him.thscp.mediator.elmis.orchestrator.ReportingTimelinessOrchestrator;
+import tz.go.moh.him.thscp.mediator.elmis.orchestrator.StockAvailabilityOrchestrator;
+import tz.go.moh.him.thscp.mediator.elmis.orchestrator.StockOnHandStatusOrchestrator;
+import tz.go.moh.him.thscp.mediator.elmis.orchestrator.TurnAroundTimeOrchestrator;
+import tz.go.moh.him.thscp.mediator.elmis.orchestrator.WorkforceOrchestrator;
 
 import java.io.File;
 import java.io.IOException;
