--- src/main/java/tz/go/moh/him/thscp/mediator/elmis/Orchestration/DefaultOrchestrator.java
@@ -1,30 +1,0 @@
-package tz.go.moh.him.thscp.mediator.elmis.Orchestration;
-
-import akka.actor.UntypedActor;
-import akka.event.Logging;
-import akka.event.LoggingAdapter;
-import org.apache.http.HttpStatus;
-import org.openhim.mediator.engine.MediatorConfig;
-import org.openhim.mediator.engine.messages.FinishRequest;
-import org.openhim.mediator.engine.messages.MediatorHTTPRequest;
-
-public class DefaultOrchestrator extends UntypedActor {
-    LoggingAdapter log = Logging.getLogger(getContext().system(), this);
-
-    private final MediatorConfig config;
-
-
-    public DefaultOrchestrator(MediatorConfig config) {
-        this.config = config;
-    }
-
-    @Override
-    public void onReceive(Object msg) throws Exception {
-        if (msg instanceof MediatorHTTPRequest) {
-            FinishRequest finishRequest = new FinishRequest("A message from my new mediator!", "text/plain", HttpStatus.SC_OK);
-            ((MediatorHTTPRequest) msg).getRequestHandler().tell(finishRequest, getSelf());
-        } else {
-            unhandled(msg);
-        }
-    }
-}
