+++ src/main/java/tz/go/moh/him/thscp/mediator/elmis/domain/ReportingTimelinessRequest.java
@@ -1,4 +1,4 @@
-package tz.go.moh.him.thscp.mediator.elmis.Domain;
+package tz.go.moh.him.thscp.mediator.elmis.domain;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.gson.annotations.SerializedName;
