+++ src/main/java/tz/go/moh/him/thscp/mediator/elmis/domain/StockAvailabilityRequest.java
@@ -1,9 +1,9 @@
-package tz.go.moh.him.thscp.mediator.elmis.Domain;
+package tz.go.moh.him.thscp.mediator.elmis.domain;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.gson.annotations.SerializedName;
 
-public class EmergencyCommodityStockStatusRequest {
+public class StockAvailabilityRequest {
     /**
      * The unique identifier
      */
@@ -12,21 +12,21 @@ public class EmergencyCommodityStockStatusRequest {
     private String uuid;
 
     /**
-     * The available quantity
+     * The facility HFR Id
      */
-    @JsonProperty("availableQuantity")
-    @SerializedName("availableQuantity")
-    private int availableQuantity;
-
-    /**
-     * The facility HFR id
-     */
-    @JsonProperty("facility_id")
-    @SerializedName("facility_id")
+    @JsonProperty("facilityId")
+    @SerializedName("facilityId")
     private String facilityId;
 
     /**
-     * The date period
+     * The incident
+     */
+    @JsonProperty("incident")
+    @SerializedName("incident")
+    private String incident;
+
+    /**
+     * The period date
      */
     @JsonProperty("period")
     @SerializedName("period")
@@ -46,20 +46,6 @@ public class EmergencyCommodityStockStatusRequest {
     @SerializedName("programCode")
     private String programCode;
 
-    /**
-     * The number of months of stock
-     */
-    @JsonProperty("stockOfMonth")
-    @SerializedName("stockOfMonth")
-    private int stockOfMonth;
-
-    /**
-     * Stock quantity
-     */
-    @JsonProperty("stockQuantity")
-    @SerializedName("stockQuantity")
-    private int stockQuantity;
-
     public String getUuid() {
         return uuid;
     }
@@ -68,20 +54,20 @@ public class EmergencyCommodityStockStatusRequest {
         this.uuid = uuid;
     }
 
-    public int getAvailableQuantity() {
-        return availableQuantity;
-    }
-
-    public void setAvailableQuantity(int availableQuantity) {
-        this.availableQuantity = availableQuantity;
-    }
-
     public String getFacilityId() {
         return facilityId;
     }
 
     public void setFacilityId(String facilityId) {
         this.facilityId = facilityId;
+    }
+
+    public String getIncident() {
+        return incident;
+    }
+
+    public void setIncident(String incident) {
+        this.incident = incident;
     }
 
     public String getPeriod() {
@@ -106,21 +92,5 @@ public class EmergencyCommodityStockStatusRequest {
 
     public void setProgramCode(String programCode) {
         this.programCode = programCode;
-    }
-
-    public int getStockOfMonth() {
-        return stockOfMonth;
-    }
-
-    public void setStockOfMonth(int stockOfMonth) {
-        this.stockOfMonth = stockOfMonth;
-    }
-
-    public int getStockQuantity() {
-        return stockQuantity;
-    }
-
-    public void setStockQuantity(int stockQuantity) {
-        this.stockQuantity = stockQuantity;
     }
 }
