+++ src/main/java/tz/go/moh/him/thscp/mediator/elmis/orchestrator/EmergencyCommodityStockStatusOrchestrator.java
@@ -1,48 +1,48 @@
-package tz.go.moh.him.thscp.mediator.elmis.Orchestration;
+package tz.go.moh.him.thscp.mediator.elmis.orchestrator;
 
 import org.codehaus.plexus.util.StringUtils;
 import org.openhim.mediator.engine.MediatorConfig;
 import org.openhim.mediator.engine.messages.MediatorHTTPRequest;
 import tz.go.moh.him.mediator.core.domain.ResultDetail;
-import tz.go.moh.him.thscp.mediator.elmis.Domain.StockOnHandStatusRequest;
+import tz.go.moh.him.thscp.mediator.elmis.domain.EmergencyCommodityStockStatusRequest;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-public class StockOnHandStatusOrchestrator extends BaseOrchestrator {
+public class EmergencyCommodityStockStatusOrchestrator extends BaseOrchestrator {
     /**
-     * Initializes a new instance of the {@link StockOnHandStatusOrchestrator} class.
+     * Initializes a new instance of the {@link EmergencyCommodityStockStatusOrchestrator} class.
      *
      * @param config The configuration.
      */
-    public StockOnHandStatusOrchestrator(MediatorConfig config) {
+    public EmergencyCommodityStockStatusOrchestrator(MediatorConfig config) {
         super(config);
     }
 
     @Override
     protected void onReceiveRequestInternal(MediatorHTTPRequest request) throws Exception {
-        List<StockOnHandStatusRequest> stockOnHandStatusRequests = Arrays.asList(serializer.deserialize(request.getBody(), StockOnHandStatusRequest[].class));
+        List<EmergencyCommodityStockStatusRequest> stockOnHandStatusRequests = Arrays.asList(serializer.deserialize(request.getBody(), EmergencyCommodityStockStatusRequest[].class));
 
         sendDataToThscp(stockOnHandStatusRequests, validateMessage(stockOnHandStatusRequests));
     }
 
     /**
-     * Validates a Stock on Hand Status request.
+     * Validates a Emergency Commodity Stock Status request.
      *
      * @param requests The requests.
      * @return Returns a list of result details.
      */
-    private List<ResultDetail> validateMessage(List<StockOnHandStatusRequest> requests) {
+    private List<ResultDetail> validateMessage(List<EmergencyCommodityStockStatusRequest> requests) {
         ArrayList<ResultDetail> results = new ArrayList<>();
 
-        for (StockOnHandStatusRequest request : requests) {
+        for (EmergencyCommodityStockStatusRequest request : requests) {
             if (StringUtils.isEmpty(request.getUuid())) {
                 results.add(new ResultDetail(ResultDetail.ResultsDetailsType.ERROR, String.format(errorMessageResource.getString("NN_ERR01"), "uuid"), null));
             }
 
             if (StringUtils.isEmpty(request.getFacilityId())) {
-                results.add(new ResultDetail(ResultDetail.ResultsDetailsType.ERROR, String.format(errorMessageResource.getString("NN_ERR01"), "facilityId"), null));
+                results.add(new ResultDetail(ResultDetail.ResultsDetailsType.ERROR, String.format(errorMessageResource.getString("NN_ERR01"), "facility_id"), null));
             }
 
             if (StringUtils.isEmpty(request.getPeriod())) {
@@ -55,10 +55,6 @@ public class StockOnHandStatusOrchestrator extends BaseOrchestrator {
 
             if (StringUtils.isEmpty(request.getProgramCode())) {
                 results.add(new ResultDetail(ResultDetail.ResultsDetailsType.ERROR, String.format(errorMessageResource.getString("NN_ERR01"), "programCode"), null));
-            }
-
-            if (StringUtils.isEmpty(request.getStockId())) {
-                results.add(new ResultDetail(ResultDetail.ResultsDetailsType.ERROR, String.format(errorMessageResource.getString("NN_ERR01"), "stockId"), null));
             }
         }
