+++ src/test/resources/mediator.properties
@@ -10,7 +10,7 @@ core.api.port=8080
 core.api.user=openhim-username
 core.api.password=openhim-password
 
-destination.host=destination-system-address
-destination.api.port=destination-system-address-port
+destination.host=localhost
+destination.api.port=8080
 destination.api.path=/destination-system-path
-destination.scheme=destination-system-scheme
+destination.scheme=http
